% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSEA.R
\name{Barplot_FeatureSetEnrichmentAnalysis}
\alias{Barplot_FeatureSetEnrichmentAnalysis}
\title{Barplot of Feature Set Enrichment Analysis results}
\usage{
Barplot_FeatureSetEnrichmentAnalysis(fsea.out, alpha = 0.05)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{alpha}{FDR threshold for calling enriched feature sets. Default is 0.05}
}
\value{
\link{ggplot} object
}
\description{
this method generates a barplot with the number of enriched feature sets per factor
}
\details{
TO-DO: IT NEEDS A BIT MORE WORK ON THE THEME
}
