% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_weights}
\alias{compare_weights}
\title{Plot the robustness of the loading matrices, ie. the definition of the factors (how they impact features), across diferent trials}
\usage{
compare_weights(models, views = "all", comparison = "all",
  show_rownames = FALSE, show_colnames = FALSE, ...)
}
\arguments{
\item{models}{a list containing \code{\link{BioFAModel}} objects.}

\item{views}{: list of views or 'all'. Correlations computed on the concatenation of the loading matrices for the given views}

\item{comparison}{tye of comparison, either 'pairwise' or 'all'}

\item{...}{extra arguments passed to pheatmap}
}
\value{
Plots a heatmap of correlation of weights columns in all models when 'comparison' is 'all'.
Otherwise, for each pair of models, a seperate heatmap is produced comparing one model againt the other.
The corresponding correlation matrix or list or pairwise correlation matrices is returned
}
\description{
Different objects of \code{\link{BioFAModel}} are compared in terms of correlation between 
their weights matrices. The correlations are computed between the columns of the concatenation of the weights matrices, only on those features which are present in all models.
Ideally, the output should look like a block diagonal matrix, suggesting that the definition of factors (how they impact features) is robust under different initialisations.
If not, it suggests that some factors are weak and not captured by all models.
}
\details{
TO-FILL
}
\references{
fill this
}
