% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Plot the robustness of the latent factors across diferent trials}
\usage{
compareModels(models, show_modelnames = FALSE)
}
\arguments{
\item{models}{a list containing \code{\link{BioFAModel}} objects.}
}
\description{
Different objects of \code{\link{BioFAModel}} are compared in terms of the final value of the ELBO statistics. 
For model selection the model with the highest ELBO value is selected.
}
