% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biofam.R
\name{create_biofam}
\alias{create_biofam}
\title{create a BioFAM object}
\usage{
create_biofam(data, samples_groups = NULL)
}
\arguments{
\item{data}{Input data can be in several formats:
\itemize{
 \item{\strong{data.frame}:}{ it requires 5 columns: sample, group, feature, view, value. 
 The "group" column indicates the condition or the experiment (a label for the samples). 
 The view indicates the assay or the -omic (a label for the features).}
 \item{\strong{Seurat object}:}{}
 \item{\strong{List of matrices}:}{ A nested list of matrices \code{Y[[i]][[j]]}. 
 The first index \code{i} for the views. 
 The second index \code{j} for the groups. 
 Sample are stored in rows and features in columns.}
 }}

\item{samples_groups}{ignore...}
}
\value{
Returns an untrained \code{\link{BioFAModel}} object
}
\description{
Method to create a \code{\link{BioFAModel}} object
}
