% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBioFAM.R
\name{createBioFAMbject}
\alias{createBioFAMbject}
\alias{createBioFAMobject}
\title{Initialize a BioFAM object}
\usage{
createBioFAMobject(data)
}
\arguments{
\item{data}{either a \code{MultiAssayExperiment} or a list of matrices with features as rows and samples as columns.}
}
\value{
Returns an untrained \code{\link{BioFAModel}} object
}
\description{
Method to initialize a \code{\link{BioFAModel}} object with a multi-omics data set.
}
\details{
If the multi-omics data is provided as a list of matrices, please make sure that features 
are stored as rows and samples are stored as columns. \cr 
If the matrices have sample names, we will use them to match the different matrices, filling the corresponding missing values. \cr
If matrices have no column names, all matrices must have the same number of columns, and you are responsible for filling any missing values.
}
