% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biofam.R
\name{.split_seurat_into_groups}
\alias{.split_seurat_into_groups}
\title{Split Seurat into groups}
\usage{
.split_seurat_into_groups(srt, samples_groups, assay = "RNA")
}
\description{
Split data in Seurat object into a list of matrices
}
