% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{factors_names}
\alias{factors_names}
\alias{factors_names<-}
\alias{factors_names,BioFAModel-method}
\alias{factors_names<-,BioFAModel,vector-method}
\title{factors_names: set and retrieve factor names}
\usage{
factors_names(object)

factors_names(object) <- value

\S4method{factors_names}{BioFAModel}(object)

\S4method{factors_names}{BioFAModel,vector}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAM}} object.}

\item{value}{a character vector of factor names}
}
\value{
character vector with the features names
}
\description{
factors_names: set and retrieve factor names
}
