% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{featureNames}
\alias{featureNames}
\alias{featureNames}
\alias{featureNames<-}
\alias{featureNames,BioFAModel-method}
\alias{featureNames<-,BioFAModel,list-method}
\title{featureNames: set and retrieve feature names}
\usage{
featureNames(object)

featureNames(object) <- value

\S4method{featureNames}{BioFAModel}(object)

\S4method{featureNames}{BioFAModel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{list of character vectors with the feature names for every view}
}
\value{
list of character vectors with the feature names for each view
}
