% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{features_names}
\alias{features_names}
\alias{features_names}
\alias{features_names<-}
\alias{features_names,BioFAModel-method}
\alias{features_names<-,BioFAModel,list-method}
\alias{features_names,BioFAModel-method}
\alias{features_names<-,BioFAModel,list-method}
\title{features_names: set and retrieve feature names}
\usage{
features_names(object)

features_names(object) <- value

\S4method{features_names}{BioFAModel}(object)

\S4method{features_names}{BioFAModel,list}(object) <- value

\S4method{features_names}{BioFAModel}(object)

\S4method{features_names}{BioFAModel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{list of character vectors with the feature names for every view}

\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{list of character vectors with the feature names for every view}
}
\value{
list of character vectors with the feature names for each view

list of character vectors with the feature names for each view
}
