% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{features_views}
\alias{features_views}
\alias{features_views,BioFAModel-method}
\alias{features_views,}
\alias{BioFAModel-method}
\title{featurs_views: retrieve feature names and views}
\usage{
features_views(object, format = "default")

\S4method{features_views}{BioFAModel}(object, format = "default")
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
data.frame with the sample names and a group for each sample
}
\description{
featurs_views: retrieve feature names and views
}
