% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_covariates}
\alias{get_covariates}
\title{get_covariates}
\usage{
get_covariates(object, covariates)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{covariates}{names of the covariates}
}
\description{
This function extracts covariates from the \code{colData} in the input \code{MultiAssayExperiment} object. \cr
Note that if you did not use \code{MultiAssayExperiment} to create your \code{\link{createBioFAMobject}}, this function will not work.
}
