% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(object, views = "all", groups = "all", features = "all",
  as.data.frame = FALSE, add_intercept = TRUE)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es). 
Default is "all".}

\item{features}{list of character vectors with the feature names or list of numeric vectors with the feature indices. 
Default is "all"}

\item{as.data.frame}{logical indicating whether to return a long data frame instead of a list of matrices.
Default is \code{FALSE}.}
}
\description{
Fetch the input data
}
\details{
By default this function returns a list where each element is a data matrix with dimensionality (D,N) 
where D is the number of features and N is the number of samples. \cr
Alternatively, if \code{as.data.frame} is \code{TRUE}, the function returns a long-formatted data frame with columns (view,feature,sample,value).
}
