% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{get_default_data_options}
\alias{get_default_data_options}
\title{Get default data options}
\usage{
get_default_data_options()
}
\value{
Returns a list with the default data options, which have to be passed as an argument to \code{\link{prepareMOFA}}
}
\description{
Function to obtain the default data options.
}
\details{
The data options are the following: \cr
\itemize{
 \item{\strong{sclae_views}:}{ logical indicating whether to scale views to have the same unit variance. 
 As long as the scale differences between the data sets is not too high, this is not required. Default is FALSE.}
 \item{\strong{center_features}:}{ logical indicating whether to center the features to zero mean. This only works for gaussian data. Default is TRUE.}
}
}
