% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{get_default_model_options}
\alias{get_default_model_options}
\title{Get default model options}
\usage{
get_default_model_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}} object}
}
\value{
Returns a list with the default model options.
}
\description{
Function to obtain the default model options.
}
\details{
The model options are the following: \cr
\itemize{
 \item{\strong{likelihoods}:}{ character vector with data likelihoods per view: 
 'gaussian' for continuous data, 'bernoulli' for binary data and 'poisson' for count data.
 By default, they are guessed internally.}
 \item{\strong{num_factors}:}{ numeric value indicating the (initial) number of factors. Default is 15.}
 \item{\strong{spikeslab_z}:}{ logical indicating whether to use spike and slab sparsity on the factors (Default is FALSE)}
 \item{\strong{spikeslab_w}:}{ logical indicating whether to use spike and slab sparsity on the weights (Default is TRUE)}
 \item{\strong{ard_z}:}{ logical indicating whether to use ARD sparsity on the factors (Default is TRUE only if using multiple groups)}
 \item{\strong{ard_w}:}{ logical indicating whether to use ARD sparsity on the weights (Default is TRUE)}
 }
}
