% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{get_default_stochastic_options}
\alias{get_default_stochastic_options}
\title{Get default stochastic options}
\usage{
get_default_stochastic_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}}}
}
\value{
Returns a list with default training options
}
\description{
Function to obtain the default options for stochastic variational inference.
}
