% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_elbo}
\alias{get_elbo}
\title{get_elbo}
\usage{
get_elbo(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\description{
Extract the value of the ELBO statistics after model training. This can be useful for model selection.
}
\details{
This can be useful for model selection.
}
