% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R, R/get_methods.R
\name{getCovariates}
\alias{getCovariates}
\alias{getCovariates}
\title{getCovariates}
\usage{
getCovariates(object, covariates)

getCovariates(object, covariates)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{covariates}{names of the covariates}

\item{object}{a \code{\link{BioFAModel}} object.}

\item{covariates}{names of the covariates}
}
\description{
This function extracts covariates from the \code{colData} in the input \code{MultiAssayExperiment} object. \cr
Note that if you did not use \code{MultiAssayExperiment} to create your \code{\link{createBioFAMobject}}, this function will not work.

This function extracts covariates from the \code{colData} in the input \code{MultiAssayExperiment} object. \cr
Note that if you did not use \code{MultiAssayExperiment} to create your \code{\link{createBioFAMobject}}, this function will not work.
}
