% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBioFAM.R
\name{getDefaultDataOptions}
\alias{getDefaultDataOptions}
\title{Get default data options}
\usage{
getDefaultDataOptions()
}
\value{
Returns a list with the default data options.
}
\description{
Function to obtain the default data options.
}
\details{
The data options are the following: \cr
\itemize{
 \item{\strong{scaleViews}:}{ logical indicating whether to scale views to have the same unit variance. 
 As long as the scale differences between the data sets is not massive, this is not required.
 Default is FALSE.}
 \item{\strong{centerFeatures}:}{ logical indicating whether to learn the intercept (the means) in a per feature basis.
  This prevents you from having to center the data, so this option is always recommended. Default is TRUE.}
 \item{\strong{removeIncompleteSamples}:}{ logical indicating whether to remove samples that are not profiled in all omics. 
  We recommend this for testing. Default is FALSE.}
 \item{\strong{delimiter}:}{ character indicating the delimiter in the input matrices stored as text files. 
  The matrices are stored internally by BioFAM, so unless you are modifying this, use the default value.}
}
}
