% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBioFAM.R
\name{getDefaultModelOptions}
\alias{getDefaultModelOptions}
\title{Get default model options}
\usage{
getDefaultModelOptions(object)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}} object}
}
\value{
Returns a list with the default model options.
}
\description{
Function to obtain the default model options.
}
\details{
The model options are the following: \cr
\itemize{
 \item{\strong{likelihood}:}{ character vector with data likelihoods per view: 
 'gaussian' for continuous data, 'bernoulli' for binary data and 'poisson' for count data.
 By default, they are guessed internally.}
 \item{\strong{numFactors}:}{ numeric indicating the initial number of factors. 
 If you want to learn the number of factors automaticallty and you have no prior expectation, 
 we recommend setting this to a large value, around 50. Default is 25.}
 \item{\strong{learnIntercept}:}{ logical indicating whether to learn an intercept term to capture differences in feature means.
 This prevents you from having to center the data, so this option is always recommended. Default is TRUE.}
 \item{\strong{sparsity}:}{ logical indicating whether to use the sparse model. 
 This is always recommended, as it will make the loadings more interpretable. Default is TRUE.}
 \item{\strong{covariates}:}{ ignore, not implemented yet.}
}
}
