% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBioFAM.R
\name{getDefaultTrainOptions}
\alias{getDefaultTrainOptions}
\title{Get default training options}
\usage{
getDefaultTrainOptions()
}
\value{
Returns a list with default training options
}
\description{
Function to obtain the default training options.
}
\details{
The training options are the following: \cr
\itemize{
 \item{\strong{maxiter}:}{ numeric value indicating the maximum number of iterations. 
 Default is 5000, but we recommend using the 'tolerance' as convergence criteria.}
 \item{\strong{tolerance}:}{ numeric value indicating the convergence threshold based on the change in Evidence Lower Bound (deltaELBO). 
 For quick exploration we recommend this to be around 1.0, and for a thorough training we recommend a value of 0.01. Default is 0.1}
 \item{\strong{learn_factors}:}{ logical indicating whether to learn the number of factors. 
 The criteria to shut down factors is based on a minimum fraction of variance explained, defined in the \code{DropFactorThreshold} option}
 \item{\strong{drop_factor_threshold}:}{ numeric indicating the threshold on fraction of variance explained to consider a factor inactive and drop it from the model.
 For example, a value of 0.01 implies that factors explaining less than 1\% of variance (in each view) will be dropped.}
 \item{\strong{verbose}:}{ logical indicating whether to generate a verbose output.}
 \item{\strong{seed}:}{ random seed for reproducibility (default is NULL, which samples a random seed).}
}
}
