% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBioFAM.R
\name{getDefaultTrainOpts}
\alias{getDefaultTrainOpts}
\title{Get default training options}
\usage{
getDefaultTrainOpts()
}
\value{
Returns a list with default training options
}
\description{
Function to obtain the default training options.
}
\details{
The training options are the following: \cr
\itemize{
 \item{\strong{maxiter}:}{ numeric indicating the maximum number of iterations. 
 Default is 1000, but we recommend using the convergence criteria.}
 \item{\strong{tolerance}:}{ numeric indicating the convergence threshold based on the change in Evidence Lower Bound (deltaELBO). 
 For quick exploration we recommend this to be around 1.0, and for a thorough training we recommend a value of 0.01}
 \item{\strong{learnFactors}:}{ logical indicating whether to learn the number of factors. 
 The criteria to shut down factors is based on a minimum fraction of variance explained, defined in the DropFactorThreshold option}
 \item{\strong{DropFactorThreshold}:}{ numeric indicating the threshold on fraction of variance explained to consider a factor inactive and drop it from the model.
 For example, a value of 0.01 implies that factors explaining less than 1\% of variance (in each view) will be dropped.}
 \item{\strong{verbose}:}{ logical indicating whether to generate a verbose output.}
}
}
