% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getDimensions}
\alias{getDimensions}
\title{getDimensions}
\usage{
getDimensions(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\description{
Extract dimensionalities from the model.
}
\details{
K indicates the number of factors, D indicates the number of features, 
N indicates the (total) number of samples and M indicates the number of views.
}
