% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getELBO}
\alias{getELBO}
\title{getELBO}
\usage{
getELBO(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\description{
Extract the value of the ELBO statistics after model training. This can be useful for model selection.
}
