% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{groupNames}
\alias{groupNames}
\alias{groupNames}
\alias{groupNames<-}
\alias{groupNames,BioFAModel-method}
\alias{groupNames<-,BioFAModel,character-method}
\title{groupNames: set and retrieve view names}
\usage{
groupNames(object)

groupNames(object) <- value

\S4method{groupNames}{BioFAModel}(object)

\S4method{groupNames}{BioFAModel,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{character vector with the names for each view}
}
\value{
character vector with the names for each view
}
