% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{groups,BioFAModel-method}
\alias{groups,}
\alias{BioFAModel-method}
\title{groups: retrieve sample names and groups}
\usage{
groups(object, format = "default")

\S4method{groups}{BioFAModel}(object, format = "default")
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
data.frame with the sample names and a group for each sample
}
\description{
groups: retrieve sample names and groups
}
