% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{groups_names}
\alias{groups_names}
\alias{groups_names<-}
\alias{groups_names,BioFAModel-method}
\alias{groups_names<-,BioFAModel,character-method}
\title{groups_names: set and retrieve view names}
\usage{
groups_names(object)

groups_names(object) <- value

\S4method{groups_names}{BioFAModel}(object)

\S4method{groups_names}{BioFAModel,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{character vector with the names for each sample group}
}
\value{
character vector with the names for each sample group
}
\description{
groups_names: set and retrieve view names
}
