% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{imputed_data}
\alias{imputed_data}
\alias{imputed_data}
\alias{imputed_data<-}
\alias{imputed_data,BioFAModel-method}
\title{imputed_data: set and retrieve imputed data}
\usage{
imputed_data(object)

.imputed_data(object) <- value

\S4method{imputed_data}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
