% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{ImputedData}
\alias{ImputedData}
\alias{ImputedData}
\alias{ImputedData<-}
\alias{ImputedData,BioFAModel-method}
\title{ImputedData: set and retrieve imputed data}
\usage{
ImputedData(object)

.ImputedData(object) <- value

\S4method{ImputedData}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
