% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values from a fitted BioFAModel}
\usage{
imputeMissing(object, views = "all", factors = "all", type = c("inRange",
  "response", "link"))
}
\arguments{
\item{object}{a \code{\link{BioFAM}} object.}

\item{views}{character vector with the view names, or numeric vector with view indexes.}

\item{factors}{character vector with the factor names, or numeric vector with the factor indexes.}

\item{type}{type of imputation. 
"response" gives mean for gaussian and poisson and probabilities for bernoulli,
"link" gives the linear predictions,
"inRange" (default) rounds the fitted values from "terms" for integer-valued distributions to the next integer.}
}
\description{
This function uses the latent factors and the loadings infered in order to impute missing values.
}
\details{
matrix factorization models generate a denoised and condensed low-dimensional representation of the data which capture the main sources of heterogeneity of the data. 
These representation can be used to do predictions using the equation \code{Y = WX}. For more details read the supplementary methods of the manuscript. \cr
This method fills the \code{ImputedData} slot by replacing the missing values in the input data with the model predictions.
}
