% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{input_data}
\alias{input_data}
\alias{input_data}
\alias{inputData<-}
\alias{input_data}
\alias{input_data}
\alias{inputData<-}
\alias{input_data}
\alias{input_data}
\alias{inputData<-}
\title{Set and retrieve input data}
\usage{
input_data(object)

.input_data(object) <- value

\S4method{input_data}{BioFAModel}(object)

\S4method{.input_data}{BioFAModel,MultiAssayExperiment}(object) <- value

\S4method{input_data}{BioFAModel}(object)

\S4method{.input_data}{BioFAModel,MultiAssayExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
