% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{InputData}
\alias{InputData}
\alias{InputData}
\alias{inputData<-}
\alias{InputData}
\alias{InputData}
\alias{inputData<-}
\title{Set and retrieve input data}
\usage{
InputData(object)

.InputData(object) <- value

\S4method{InputData}{BioFAModel}(object)

\S4method{.InputData}{BioFAModel,MultiAssayExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
