% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSEA.R
\name{LinePlot_FeatureSetEnrichmentAnalysis}
\alias{LinePlot_FeatureSetEnrichmentAnalysis}
\title{Line plot of Feature Set Enrichment Analysis results}
\usage{
LinePlot_FeatureSetEnrichmentAnalysis(fsea.out, factor, threshold = 0.1,
  max.pathways = 25, adjust = T)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{factor}{Factor for which to show wnriched pathways in the lineplot}

\item{threshold}{p.value threshold to filter out feature sets}

\item{max.pathways}{maximum number of enriched pathways to display}

\item{adjust}{use multiple testing correction}
}
\value{
nothing
}
\description{
Line plot of the Feature Set Enrichment Analyisis results for a specific latent variable
}
\details{
fill this
}
