% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R, R/load_model2.R
\name{load_model}
\alias{load_model}
\alias{load_model}
\alias{load_model2}
\title{loading a trained BioFAModel}
\usage{
load_model(file, object = NULL, sort_factors = TRUE)

load_model2(reticulate, object = NULL, sort_factors = TRUE,
  multi_view = NULL, multi_group = NULL)
}
\arguments{
\item{file}{TO-FILL}

\item{object}{TO-FILL}

\item{sort_factors}{TO-FILL}
}
\value{
a \code{\link{BioFAModel}} model

a \code{\link{BioFAModel}} model
}
\description{
TO-FILL

Method to load a trained BioFAModel
}
