% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{loading a trained BioFAModel}
\usage{
loadModel(file, object = NULL, sortFactors = T)
}
\arguments{
\item{file}{an hdf5 file saved by the BioFAM python framework.}

\item{object}{either NULL (default) or an an existing untrained BioFAM object. If NULL, the \code{\link{BioFAModel}} object is created from the scratch.}

\item{sortFactors}{boolean indicating whether factors should be sorted by variance explained (default is TRUE)}
}
\value{
a \code{\link{BioFAModel}} model.
}
\description{
Method to load a trained BioFAModel. \cr
The training of BioFAM is done using a Python framework, and the model output is saved as an .hdf5 file, which has to be loaded in the R package.
}
