% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{ModelOptions}
\alias{ModelOptions}
\alias{ModelOptions}
\alias{ModelOptions<-}
\alias{ModelOptions,BioFAModel-method}
\title{ModelOptions: set and retrieve Model options}
\usage{
ModelOptions(object)

.ModelOptions(object) <- value

\S4method{ModelOptions}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of model options
}
