% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{ModelOpts}
\alias{ModelOpts}
\alias{ModelOpts}
\alias{ModelOpts<-}
\alias{ModelOpts,BioFAModel-method}
\title{ModelOpts: set and retrieve Model options}
\usage{
ModelOpts(object)

.ModelOpts(object) <- value

\S4method{ModelOpts}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of model options
}
