% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data_overview}
\alias{plot_data_overview}
\title{Tile plot of the multi-omics data}
\usage{
plot_data_overview(object, colors = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{colors}{a vector specifying the colors per view.}
}
\description{
Function to do a tile plot showing the missing value structure of the multi-omics input data
}
\details{
This function is helpful to get an overview of the missing value structure of the training data used for BioFAM. 
It shows the number of samples, the number of views, the number of features, and the structure of missing values.
In particular, it is useful to visualise incomplete data sets, where some samples are missing subsets of assays.
}
