% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data_scatter}
\alias{plot_data_scatter}
\title{Scatterplots of feature values against latent factors}
\usage{
plot_data_scatter(object, view, factor, groups = "all", features = 10,
  sign = "all", color_by = NULL, color_name = "",
  color_legend = TRUE, shape_by = NULL, shape_name = "",
  shape_legend = TRUE, dot_size = 1, text_size = 5, add_lm = TRUE,
  imputed = FALSE)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{view}{character vector with a view name, or numeric vector with the index of the view.}

\item{factor}{character vector with a factor name, or numeric vector with the index of the factor.}

\item{groups}{TO-FILL}

\item{features}{if an integer, the total number of features to plot (10 by default). If a character vector, a set of manually-defined features.}

\item{sign}{can be 'positive', 'negative' or 'all' to show only positive, negative or all weights, respectively. Default is 'all'.}

\item{color_by}{specifies groups or values (either discrete or continuous) used to color the dots (samples). This can be either: 
\itemize{
\item (default) the string "group", it the dots with respect to their predefined groups.
\item a character giving the name of a feature that is present in the input data 
\item a character giving the same of a column in the sample metadata slot
\item a vector of the same length as the number of samples specifying the value for each sample. 
\item a dataframe with two columns: "sample" and "color"
}}

\item{color_name}{name for color legend (usually only used if color_by is not a character itself).}

\item{shape_by}{specifies groups or values (only discrete) used to shape the dots (samples). This can be either: 
\itemize{
\item (default) the string "group": in this case, the plot will shape the dots with respect to their predefined groups.
\item a character giving the name of a feature that is present in the input data 
\item a character giving the same of a column in the sample metadata slot
\item a vector of the same length as the number of samples specifying the value for each sample. 
\item a dataframe with two columns: "sample" and "shape"
}}

\item{shape_name}{name for shape legend (usually only used if shape_by is not a character itself).}

\item{dot_size}{numeric indicating dot size.}

\item{text_size}{numeric indicating text size.}

\item{add_lm}{logical indicating whether to add a linear regression line for each plot}

\item{imputed}{logical indicating whether to include imputed measurements}
}
\description{
Function to do a scatterplot of the feature(s) values against the latent factor values.
}
\details{
One of the first steps for the annotation of factors is to visualise the loadings using \code{\link{plot_weights}} or \code{\link{plot_top_weights}}, 
which show you which features drive the heterogeneity of each factor. 
However, one might also be interested in visualising the direct relationship between features and factors, rather than looking at "abstract" weights. \cr
This function generates scatterplots of features against factors, so that you can observe the association between them. \cr
A similar function for doing heatmaps rather than scatterplots is \code{\link{plot_data_heatmap}}.
}
