% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factor}
\alias{plot_factor}
\title{Beeswarm plots of factor values}
\usage{
plot_factor(object, factors = "all", group_by = "group",
  add_dots = TRUE, add_violin = TRUE, show_missing = TRUE,
  dot_size = 1, color_by = NULL, color_name = "", shape_by = NULL,
  alpha = 0.25, shape_name = "", rasterize = FALSE, dodge = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. 
Default is 'all'}

\item{group_by}{specifies groups used to separate the samples : one plot per group 
This can be either: 
the string "group" : in this case, the plot will separate samples with respect to the groups they belong to
a character giving the name of a feature, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.
or a dataframe with two columns : "sample" (names of the samples) and "group" (groups for each sample)}

\item{show_missing}{logical indicating whether to remove samples for which \code{shape_by} or \code{color_by} is missing.}

\item{color_by}{specifies groups or values (discrete or continuous) used to color the samples
This can be either: 
the string "group" : in this case, the plot will color samples with respect to the groups they belong to
a character giving the name of a feature, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.
or a dataframe with two columns : "sample" (names of the samples) and "color" (values for each sample)}

\item{color_name}{name for color legend (usually only used if color_by is not a character itself)}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Beeswarm plot of the latent factor values.
}
\details{
One of the main steps for the annotation of factors is to visualise and color them using known covariates or phenotypic data. \cr
This function generates a Beeswarm plot of the sample values in a given latent factor. \cr
Similar functions are \code{\link{plot_factor_scatter}} for doing scatter plots.
}
