% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factor_scatters}
\alias{plot_factor_scatters}
\title{Pairwise scatterplots of multiple latent factors}
\usage{
plot_factor_scatters(object, factors = "all", groups = "all",
  show_missing = TRUE, color_by = NULL, name_color = "",
  shape_by = NULL, name_shape = "")
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. 
Default is 'all'}

\item{show_missing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either: 
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either: 
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{name_shape}{name for shape legend (usually only used if shape_by is not a character itself)}
}
\value{
\code{ggplot2} object
}
\description{
Scatterplots of the sample values for pair-wise combinations of multiple latent factors.
}
\details{
One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates multiple scatterplots for pairwise combinations of several latent factors.
Similar functions are \code{\link{plot_factor_scatter}} for doing single scatter plots and 
\code{\link{plot_factor_beeswarm}} for doing Beeswarm plots for single factors.
}
