% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_summary.R
\name{plot_factor_summary}
\alias{plot_factor_summary}
\title{Plot summary of informative plots relative to a specific factor}
\description{
Plot summary of informative plots relative to a specific factor
}
