% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\alias{plot_factor}
\title{Beeswarm plots of factor values}
\usage{
plot_factor(object, factor = 1, group_by = "group", add_dots = TRUE,
  add_violin = FALSE, show_missing = TRUE, dot_size = 1,
  scale = FALSE, color_by = NULL, color_name = "", shape_by = NULL,
  shape_name = "", dots_alpha = 1, legend = TRUE,
  violin_alpha = 0.5, color_violin = TRUE, rasterize = FALSE,
  dodge = FALSE)

plot_factors(object, factors = c(1, 2), show_missing = TRUE,
  scale = FALSE, color_by = NULL, shape_by = NULL,
  color_name = NULL, shape_name = NULL, dot_size = 1.5, alpha = 1,
  legend = TRUE, return_data = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{group_by}{specifies groups used to separate the samples : one plot per group 
This can be either: 
the string "group" : in this case, the plot will separate samples with respect to the groups they belong to
a character giving the name of a feature, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.
or a dataframe with two columns : "sample" (names of the samples) and "group" (groups for each sample)}

\item{show_missing}{logical indicating whether to remove samples for which \code{shape_by} or \code{color_by} is missing.}

\item{color_by}{specifies groups or values (discrete or continuous) used to color the samples
This can be either: 
the string "group" : in this case, the plot will color samples with respect to the groups they belong to
a character giving the name of a feature, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.
or a dataframe with two columns : "sample" (names of the samples) and "color" (values for each sample)}

\item{color_name}{name for color legend (usually only used if color_by is not a character itself)}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{shape_name}{name for shape legend (usually only used if shape_by is not a character itself)}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. 
Default is 'all'}

\item{return_data}{logical indicating whether to return the data frame to plot instead of plotting}

\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{a vector of length two with the factors to plot. Factors can be specified either as a characters
using the factor names, or as numeric with the index of the factors}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either 
the string "group" : in this case, the plot will color samples with respect to the groups they belong to
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{color_name}{name for color legend (usually only used if color_by is not a character itself)}

\item{show_missing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}
}
\value{
Returns a \code{ggplot2} object

Returns a \code{ggplot2} object
}
\description{
Beeswarm plot of the latent factor values.

Scatterplot of the values of two latent factors.
}
\details{
One of the main steps for the annotation of factors is to visualise and color them using known covariates or phenotypic data. \cr
This function generates a Beeswarm plot of the sample values in a given latent factor. \cr
Similar functions are \code{\link{plot_factor_scatter}} for doing scatter plots.

One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a single scatterplot for the combination of two latent factors.
Similar function is
\code{\link{plot_factors}} for doing Beeswarm plots for factors.
}
