% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_training_stats.R
\name{plot_training_curve_factors}
\alias{plot_training_curve_factors}
\title{Training curve for the number of active factors}
\usage{
plot_training_curve_factors(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\description{
the BioFAModel starts with an initial number of factors and inactive factors are dropped during training if they explain small amounts of variation. 
This allows the model to automatically infer the dimensionality of the latent space.
The corresponding hyperparameters are defined in \code{\link{prepareBioFAM}}. \cr
All training statistics, including the number of active factors, can be fetch from the training_stats slot of \code{\link{BioFAModel}} .
}
