% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables_factors_corrs.R
\name{plot_variables_factors_corrs}
\alias{plot_variables_factors_corrs}
\title{Plot weights and correlations bewteen factors and expressions of the variables in each view for both groups.}
\arguments{
\item{variables}{: list of features names of the biofam model, or named vector where names are names of variables (covariates) and values are named vectors storing the values of the variables for each sample}

\item{viz_weight}{: option relevant only if variables are features names of the biofam model. If true, plot weights and not correlation across all groups in the first figures}
}
