% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_explained.R
\name{plot_variance_explained_in_groups}
\alias{plot_variance_explained_in_groups}
\title{Plot variance explained by the model in groups}
\usage{
plot_variance_explained_in_groups(object, views = "all", cluster = T, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{cluster}{logical indicating whether to do hierarchical clustering on the plot}

\item{...}{extra arguments to be passed to \code{\link{calculate_variance_explained}}}
}
\value{
ggplot object
}
\description{
Method to plot variance explained (R-squared) by the MOFA model for each group and latent factor. \cr
As a measure of variance explained for gaussian data we adopt the coefficient of determination (R2). \cr
For details on the computation see the help of the \code{\link{calculate_variance_explained}} function
}
