% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot_weight_scatter}
\alias{plot_weight_scatter}
\title{Scatterplot of weights for two factors}
\usage{
plot_weight_scatter(object, view, factors, color_by = NULL, shape_by = NULL,
  name_color = "", name_shape = "", showMissing = TRUE)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{view}{character vector with the voiew name, or numeric vector with the index of the view to use.}

\item{factors}{a vector of length two with the factors to plot. Factors can be specified either as a characters
using the factor names, or as numeric with the index of the factors}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either 
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{name_shape}{name for shape legend (usually only used if shape_by is not a character itself)}

\item{showMissing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Scatterplot of the weights values for two factors
}
\details{
One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a single scatterplot for the combination of two latent factors.
}
