% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot Weights}
\usage{
plot_weights(object, views = "all", factors = "all", nfeatures = 10,
  abs = FALSE, manual = NULL, color_manual = NULL, scale = TRUE,
  sort_by_factor = "all", dot_size = 1, text_size = 5)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{views}{character vector with the voiew name, or numeric vector with the index of the view to use.}

\item{factors}{character vector with the factor name, or numeric vector with the index of the factor to use.}

\item{nfeatures}{number of top features to label.}

\item{abs}{logical indicating whether to use the absolute value of the weights.}

\item{manual}{A nested list of character vectors with features to be manually labelled.}

\item{color_manual}{a character vector with colors, one for each element of 'manual'}

\item{scale}{logical indicating whether to scale all loadings from 0 to 1.}

\item{sort_by_factor}{name or numeric value for the factor to sort all loadings by, or all (default)
when "all", features between factors do not match}
}
\description{
An important step to annotate factors is to visualise the corresponding feature loadings. \cr
This function plots all loadings for a given latent factor and view, labeling the top ones. \cr
In contrast, the function \code{\link{plot_top_weights}} displays only the top features with highest loading.
}
\details{
The weights of the features within a view are relative andthey should not be interpreted in an absolute scale.
Therefore, for interpretability purposes we always recommend to scale the weights with \code{scale=TRUE}.
}
