% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_weights_signature}
\alias{plot_weights_signature}
\title{Barchart of feature loadings}
\usage{
plot_weights_signature(object, factors = "all", n_top = 10, ...)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. 
Default is 'all'}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Barchar of the feature loadings.
}
