% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R, R/plot_factors.R
\name{plotFactorScatter}
\alias{plotFactorScatter}
\alias{plotFactorScatter}
\title{Scatterplot of two latent factors}
\usage{
plotFactorScatter(object, factors, color_by = NULL, shape_by = NULL,
  name_color = "", name_shape = "", showMissing = TRUE)

plotFactorScatter(object, factors, color_by = NULL, shape_by = NULL,
  name_color = "", name_shape = "", showMissing = TRUE)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{a vector of length two with the factors to plot. Factors can be specified either as a characters
using the factor names, or as numeric with the index of the factors}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either 
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{name_shape}{name for shape legend (usually only used if shape_by is not a character itself)}

\item{showMissing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}

\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{a vector of length two with the factors to plot. Factors can be specified either as a characters
using the factor names, or as numeric with the index of the factors}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either 
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{name_shape}{name for shape legend (usually only used if shape_by is not a character itself)}

\item{showMissing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}
}
\value{
Returns a \code{ggplot2} object

Returns a \code{ggplot2} object
}
\description{
Scatterplot of the values of two latent factors.

Scatterplot of the values of two latent factors.
}
\details{
One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a single scatterplot for the combination of two latent factors.
Similar functions are \code{\link{plotFactorScatters}} for doing multiple scatter plots and 
\code{\link{plotFactorBeeswarm}} for doing Beeswarm plots for single factors.

One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a single scatterplot for the combination of two latent factors.
Similar functions are \code{\link{plotFactorScatters}} for doing multiple scatter plots and 
\code{\link{plotFactorBeeswarm}} for doing Beeswarm plots for single factors.
}
