% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFactors.R, R/plot_factors.R
\name{plotSparsityFactors}
\alias{plotSparsityFactors}
\alias{plotSparsityFactors}
\title{Plot the sparsity of factors}
\usage{
plotSparsityFactors(object, threshold_variance_explained = 0.01,
  show_variance_explained = T)

plotSparsityFactors(object, threshold_variance_explained = 0.01,
  show_variance_explained = T)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{threshold_variance_explained}{: do not plot the factors explaining less of the variance explained than the threshold for the corresponding views}

\item{show_variance_explained}{: indicate the variance explained by each of the factor in each view}

\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{threshold_variance_explained}{: do not plot the factors explaining less of the variance explained than the threshold for the corresponding views}

\item{show_variance_explained}{: indicate the variance explained by each of the factor in each view}
}
\value{
Returns a \code{ggplot2} object

Returns a \code{ggplot2} object
}
\description{
Plot the sparsity of the factors (computed as 1-<Theta_k> for factor k)

Plot the sparsity of the factors (computed as 1-<Theta_k> for factor k)
}
\details{
This method enables to visualize simultaneously the factors that are active in the different views (with the use of the threshold)
and their sparsity level. A sparse factor could be biologically meaningfull, while a dense one is probably a technical source of variability...

This method enables to visualize simultaneously the factors that are active in the different views (with the use of the threshold)
and their sparsity level. A sparse factor could be biologically meaningfull, while a dense one is probably a technical source of variability...
}
