% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Do predictions using a fitted BioFAModel}
\usage{
predict(object, views = "all", groups = "all", factors = "all",
  type = c("inRange", "response", "link"), include_intercept = TRUE)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es). 
Default is "all".}

\item{factors}{character vector with the factor name(s) or numeric vector with the factor index(es). 
Default is "all".}

\item{type}{type of prediction returned, either: 
"response" gives the response vector, the mean for Gaussian and Poisson, and probabilities for Bernoulli, 
"link" gives the linear predictions, 
"inRange" rounds the fitted values integer-valued distributions to the next integer. \cr
Default is "inRange".}

\item{include_intercept}{logical indicating whether to include the intercept factors for the prediction, if present.
Default is TRUE.}
}
\value{
Returns a list with data predictions, each element corresponding to a view.
}
\description{
This function uses the latent factors and the weights to do data predictions.
}
\details{
the denoised and condensed low-dimensional representation of the data captures the main sources of heterogeneity of the data. 
These representation can be used to do predictions using the equation Y = WX. This is the key step underlying imputation, see \code{\link{imputeMissing}} and Methods section of the article.
}
