% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{prepare_biofam}
\alias{prepare_biofam}
\title{prepare a BioFAModel for training}
\usage{
prepare_biofam(object, data_options = NULL, model_options = NULL,
  training_options = NULL, center = TRUE, regress_covariates = NULL)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}}}

\item{data_options}{list of data_options (see \code{\link{get_default_data_options}} details). 
If NULL, default data options are used.}

\item{model_options}{list of model options (see \code{\link{get_default_model_options}} for details). 
If NULL, default model options are used.}

\item{training_options}{list of training options (see \code{\link{get_default_training_options}} for details). 
If NULL, default training options are used.}

\item{dir_options}{list with Input/Output options, which must contain a 'data_dir' element where temporary text files will be stored 
and an 'outFile' with hdf5 extension where the final model will be stored.}
}
\value{
Returns an untrained \code{\link{BioFAModel}} with specified data, model and training options
}
\description{
Function to prepare a \code{\link{BioFAModel}} object for training. 
It requires defining data, model and training options.
}
