% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biofam.R
\name{run_biofam}
\alias{run_biofam}
\title{train a BioFAModel}
\usage{
run_biofam(object, outfile = NULL)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}} object}

\item{outfile}{output .hdf5 file}
}
\value{
a trained \code{\link{BioFAModel}} object
}
\description{
Function to train an untrained \code{\link{BioFAModel}} object.
}
\details{
In this step the R package is calling the \code{biofam} Python package,
 where the the training is performed. \cr
The interface with Python is done with the \code{\link{reticulate}} package. 
If you have several versions of Python installed and Rstudio is not detecting
 the correct one, you can change it using' \code{reticulate::use_python}.
}
