% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sampleNames,BioFAModel-method}
\alias{sampleNames<-,BioFAModel,list-method}
\title{sampleNames: set and retrieve sample names}
\usage{
sampleNames(object)

sampleNames(object) <- value

\S4method{sampleNames}{BioFAModel}(object)

\S4method{sampleNames}{BioFAModel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{list of character vectors with the sample names for every group}
}
\value{
list of character vectors with the sample names for each group
}
