% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{samples_groups}
\alias{samples_groups}
\alias{samples_groups,BioFAModel-method}
\alias{samples_groups,}
\alias{BioFAModel-method}
\title{samples_groups: retrieve sample names and groups}
\usage{
samples_groups(object)

\S4method{samples_groups}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
data.frame with the sample names and a group for each sample
}
