% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{samples_metadata}
\alias{samples_metadata}
\alias{samples_metadata<-}
\alias{samples_metadata,BioFAModel-method}
\alias{samples_metadata<-,BioFAModel,data.frame-method}
\title{samples_metadata: retrieve sample metadata}
\usage{
samples_metadata(object, format = "default")

samples_metadata(object) <- value

\S4method{samples_metadata}{BioFAModel}(object)

\S4method{samples_metadata}{BioFAModel,data.frame}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{data frame with sample information, has to contain columns sample_name and group_name}
}
\value{
a data frame with sample metadata
}
\description{
samples_metadata: retrieve sample metadata
}
