% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{samples_names}
\alias{samples_names}
\alias{samples_names<-}
\alias{samples_names,BioFAModel-method}
\alias{samples_names<-,BioFAModel,list-method}
\title{samples_names: set and retrieve sample names}
\usage{
samples_names(object)

samples_names(object) <- value

\S4method{samples_names}{BioFAModel}(object)

\S4method{samples_names}{BioFAModel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{list of character vectors with the sample names for every group}
}
\value{
list of character vectors with the sample names for each group
}
\description{
samples_names: set and retrieve sample names
}
