% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{select_model}
\alias{select_model}
\alias{selectModel}
\title{Select a model from a list of trained \code{\link{BioFAModel}} objects based on the best ELBO value}
\usage{
selectModel(models, plotit = TRUE)
}
\arguments{
\item{models}{a list containing \code{\link{BioFAModel}} objects.}
}
\description{
Different objects of \code{\link{BioFAModel}} are compared in terms of the final value of the ELBO statistics 
and the model with the highest ELBO value is selected.
}
