% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetFactors}
\alias{subsetFactors}
\title{Subset factors}
\usage{
subsetFactors(object, factors)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{factors}{character vector with the factor names, or numeric vector with the index of the factors.}
}
\description{
Method to subset (or sort) factors
}
