% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainData}
\alias{TrainData}
\alias{TrainData}
\alias{TrainData<-}
\alias{TrainData,BioFAModel-method}
\title{TrainData: set and retrieve training data}
\usage{
TrainData(object)

.TrainData(object) <- value

\S4method{TrainData}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
