% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{training_options}
\alias{training_options}
\alias{training_options}
\alias{training_options<-}
\title{training_options: set and retrieve training opts}
\usage{
training_options(object)

.training_options(object) <- value
}
