% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{training_stats}
\alias{training_stats}
\alias{training_stats}
\alias{training_stats<-}
\alias{training_stats,BioFAModel-method}
\alias{training_stats,BioFAModel-method}
\title{training_stats: set and retrieve training statistics}
\usage{
training_stats(object)

.training_stats(object) <- value

\S4method{training_stats}{BioFAModel}(object)

\S4method{training_stats}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of training statistics

list of training statistics
}
