% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainOptions}
\alias{TrainOptions}
\alias{TrainOptions}
\alias{TrainOptions<-}
\alias{TrainOptions,BioFAModel-method}
\title{TrainOptions: set and retrieve training opts}
\usage{
TrainOptions(object)

.TrainOptions(object) <- value

\S4method{TrainOptions}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of training options
}
