% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainOpts}
\alias{TrainOpts}
\alias{TrainOpts}
\alias{TrainOpts<-}
\alias{TrainOpts,BioFAModel-method}
\title{TrainOpts: set and retrieve training opts}
\usage{
TrainOpts(object)

.TrainOpts(object) <- value

\S4method{TrainOpts}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of training options
}
