% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{TrainStats}
\alias{TrainStats}
\alias{TrainStats}
\alias{TrainStats<-}
\alias{TrainStats,BioFAModel-method}
\title{TrainStats: set and retrieve training statistics}
\usage{
TrainStats(object)

.TrainStats(object) <- value

\S4method{TrainStats}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of training statistics
}
