% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R
\docType{methods}
\name{viewNames}
\alias{viewNames}
\alias{viewNames}
\alias{viewNames<-}
\alias{viewNames,BioFAModel-method}
\alias{viewNames<-,BioFAModel,character-method}
\title{viewNames: set and retrieve view names}
\usage{
viewNames(object)

viewNames(object) <- value

\S4method{viewNames}{BioFAModel}(object)

\S4method{viewNames}{BioFAModel,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{character vector with the names for each view}
}
\value{
character vector with the names for each view
}
