% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{views_names}
\alias{views_names}
\alias{views_names}
\alias{views_names<-}
\alias{views_names,BioFAModel-method}
\alias{views_names<-,BioFAModel,character-method}
\title{views_names: set and retrieve view names}
\usage{
views_names(object)

views_names(object) <- value

\S4method{views_names}{BioFAModel}(object)

\S4method{views_names}{BioFAModel,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{value}{character vector with the names for each view}
}
\value{
character vector with the names for each view
}
