% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BioFAModel}
\alias{BioFAModel}
\alias{BioFAModel-class}
\title{Class to store a biofam model}
\description{
The \code{BioFAModel} is an S4 class used to store all relevant data to analyse a BioFAModel
}
\section{Slots}{

 \itemize{
   \item{\code{input_data}:}{ TO SPECIFY, the input data before being parsed to Training Data. 
   }
   \item{\code{training_data}:}{ TO SPECIFY the parsed data used to fit the BioFAModel 
   }
   \item{\code{imputed_data}:}{ TO SPECIFY the parsed data with the missing values imputed using the BioFAModel 
   }
   \item{\code{expectations}:}{ expected values of the different nodes of the model. }
   
   \item{\code{training_stats}:}{ list with training statistics such as evidence lower bound (ELBO) and number of factors per iteration}
   \item{\code{data_options}:}{ list with the data processing options such as whether to center or scale the data.}
   \item{\code{training_options}:}{ list with the training options such as maximum number of iterations, tolerance for convergence, etc.}
   \item{\code{model_options}:}{ list with the model options such as likelihoods, whether the intercept was learnt, etc.}
   \item{\code{dimensions}:}{ list with the relevant dimensionalities of the model. 
   N for the number of samples, 
   P for the number of sample groups, 
   D for the number of features of each view,
   M for the number of views, 
   K for the number of factors.}
   \item{\code{status}:}{Auxiliary variable indicating whether the model has been trained.}
}
}

