% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_biofam.R
\name{create_biofam}
\alias{create_biofam}
\title{create a BioFAM object}
\usage{
create_biofam(data)
}
\arguments{
\item{data}{TO SPECIFY}
}
\value{
Returns an untrained \code{\link{BioFAModel}} object
}
\description{
Method to create a \code{\link{BioFAModel}} object
}
