% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{expectations}
\alias{expectations}
\alias{expectations}
\alias{expectations<-}
\alias{expectations,BioFAModel-method}
\alias{expectations,BioFAModel-method}
\title{expectations: set and retrieve expectations}
\usage{
expectations(object)

.expectations(object) <- value

\S4method{expectations}{BioFAModel}(object)

\S4method{expectations}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of expectations

list of expectations
}
