% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{get_default_model_options}
\alias{get_default_model_options}
\title{Get default model options}
\usage{
get_default_model_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}} object}
}
\value{
Returns a list with the default model options, which have to be passed as an argument to \code{\link{prepareMOFA}}
}
\description{
Function to obtain the default model options.
}
