% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_biofam.R
\name{get_default_stochastic_options}
\alias{get_default_stochastic_options}
\title{Get default stochastic options}
\usage{
get_default_stochastic_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{BioFAModel}}}
}
\value{
Returns a list with default options
}
\description{
Function to obtain the default options for stochastic variational inference.
}
\details{
The training options are the following: \cr
\itemize{
 \item{\strong{batch_size}:}{ numeric value indicating the batch size (as a fraction)}. 
 Default is 0.5 (half of the data set).
 \item{\strong{learning_rate}:}{ numeric value indicating the learning rate. }
 Default is 0.75 
 \item{\strong{forgetting_rate}:}{ numeric indicating the forgetting rate.}
 Default is 1.0
 }
}
