% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R, R/get_methods.R
\name{getELBO}
\alias{getELBO}
\alias{getELBO}
\title{getELBO}
\usage{
getELBO(object)

getELBO(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
\description{
Extract the value of the ELBO statistics after model training. This can be useful for model selection.

Extract the value of the ELBO statistics after model training. This can be useful for model selection.
}
