% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{loading a trained BioFAModel}
\usage{
load_model(file, object = NULL, sort_factors = TRUE, on_disk = FALSE)
}
\arguments{
\item{file}{an hdf5 file saved by the biofam Python framework}

\item{object}{either NULL (default) or an an existing untrained biofam object. If NULL, the \code{\link{BioFAModel}} object is created from the scratch.}

\item{sort_factors}{boolean indicating whether factors should be sorted by variance explained (default is TRUE)}

\item{on_disk}{boolean indicating whether to work from memory (FALSE) or disk (TRUE)}
}
\value{
a \code{\link{BioFAModel}} model
}
\description{
Method to load a trained BioFAModel \cr
The training of biofam is done using a Python framework, and the model output is saved as an .hdf5 file, which has to be loaded in the R package.
}
