% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{model_options}
\alias{model_options}
\alias{model_options}
\alias{model_options<-}
\alias{model_options,BioFAModel-method}
\alias{model_options,BioFAModel-method}
\title{model_options: set and retrieve Model options}
\usage{
model_options(object)

.model_options(object) <- value

\S4method{model_options}{BioFAModel}(object)

\S4method{model_options}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of model options

list of model options
}
