% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factor_beeswarm}
\alias{plot_factor_beeswarm}
\title{Beeswarm plot of latent factors}
\usage{
plot_factor_beeswarm(object, factors = "all", color_by = NULL,
  superimpose_groups = TRUE, name_color = "", show_missing = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s) to use. 
Default is 'all'}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either: 
the string "group" : in this case, the plot will color samples with respect to the groups they belong to
a character giving the name of a feature, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.
or a dataframe with two columns : "sample" (names of the samples) and "color_by" (values of the groups)}

\item{superimpose_groups}{: if FALSE, plot the values for the factors expressions in one graph per group}

\item{name_color}{name for color legend (usually only used if color_by is not a character itself)}

\item{show_missing}{logical indicating whether to remove samples for which \code{shape_by} or \code{color_by} is missing.}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Beeswarm plot of the latent factor values.
}
\details{
One of the main steps for the annotation of factors is to visualise and color them using known covariates or phenotypic data. \cr
This function generates a Beeswarm plot of the sample values in a given latent factor. \cr
Similar functions are \code{\link{plot_factor_scatter}} for doing scatter plots and \code{\link{plot_factor_hist}} for doing histogram plots
}
