% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot_top_weights}
\alias{plot_top_weights}
\title{Plot top weights}
\usage{
plot_top_weights(object, view, factor, nfeatures = 10, abs = TRUE,
  scale = TRUE, sign = "all")
}
\arguments{
\item{object}{a trained \code{\link{BioFAModel}} object.}

\item{view}{a string with the view name, or an integer with the index of the view.}

\item{nfeatures}{number of top features to display.
Default is 10}

\item{abs}{logical indicating whether to use the absolute value of the weights (Default is FALSE).}

\item{scale}{logical indicating whether to scale all loadings from -1 to 1 (or from 0 to 1 if abs=TRUE). Default is TRUE.}

\item{sign}{can be 'positive', 'negative' or 'all' to show only positive, negative or all weights, respectively. Default is 'all'.}

\item{factors}{a string with the factor name, or an integer with the index of the factor.}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Plot top weights for a given factor and view.
}
\details{
An important step to annotate factors is to visualise the corresponding feature loadings. \cr
This function displays the top features with highest loading whereas the function \code{\link{plot_top_weights}} plots all loadings for a given latent factor and view. \cr
Importantly, the weights of the features within a view have relative values and they should not be interpreted in an absolute scale.
Therefore, for interpretability purposes we always recommend to scale the weights with \code{scale=TRUE}.
}
