% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_explained.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot variance explained by the model

Returns a list of plots with specifies axes.

Consider using cowplot::plot_grid(plotlist = ...) in order to combine plots.}
\usage{
plot_variance_explained(object, x = "view", y = "factor",
  split_by = NA, cluster = TRUE, plot_total = FALSE,
  factors = "all", legend = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object}

\item{x}{string specifying the X axis (view, factor, or group)}

\item{y}{string specifying the Y axis (view, factor, or group)}

\item{split_by}{string specifying the dimension to split a plot by (view, factor, or group)}

\item{cluster}{logical value indicating whether to do hierarchical clustering on the plot}

\item{plot_total}{logical value to indicate if to plot the total variance explained along the X axis}

\item{...}{extra arguments to be passed to \code{\link{calculate_variance_explained}}}
}
\value{
ggplot object
}
\description{
Plot variance explained by the model

Returns a list of plots with specifies axes.

Consider using cowplot::plot_grid(plotlist = ...) in order to combine plots.
}
