% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_explained.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot variance explained by the model}
\usage{
plot_variance_explained(object, cluster = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{cluster}{logical indicating whether to do hierarchical clustering on the plot}

\item{...}{extra arguments to be passed to \code{\link{calculate_variance_explained}}}
}
\value{
ggplot object
}
