% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_explained.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot variance explained by the model

Returns a list of plots with specifies axes.

Consider using cowplot::plot_grid(plotlist = ...) in order to combine plots.}
\usage{
plot_variance_explained(object, x = "view", y = "factor",
  split_by = NA, plot_total = FALSE, factors = "all", min_r2 = 0,
  legend = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object}

\item{x}{string specifying the x-axis ("view", "factor", or "group")}

\item{y}{string specifying the y-axis ("view", "factor", or "group")}

\item{split_by}{string specifying the dimension to be faceted ("view", "factor", or "group")}

\item{plot_total}{logical value to indicate if to plot the total variance explained (for the variable in the x-axis)}

\item{factors}{TO-FILL. Default is "all"}

\item{legend}{logical indicating whether to add a legend}

\item{...}{extra arguments to be passed to \code{\link{calculate_variance_explained}}}

\item{cluster}{logical value indicating whether to do hierarchical clustering on the plot}
}
\value{
ggplot object
}
\description{
Plot variance explained by the model

Returns a list of plots with specifies axes.

Consider using cowplot::plot_grid(plotlist = ...) in order to combine plots.
}
