% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biofam.R
\name{run_biofam}
\alias{run_biofam}
\title{train a BioFAM model}
\usage{
run_biofam(object, dir_options)
}
\arguments{
\item{object}{an untrained \code{\link{bioFAMmodel}} object}

\item{dir_options}{list with I/O options, it should contain at least two entries: \cr
\itemize{
 \item{\strong{data_dir}:}{ input directory where the matrices as stored as text files with row and column names and tab delimiter. }
 \item{\strong{outfile}:}{ output file where the model is going to be stored as an hdf5 file.}
}}
}
\value{
a trained \code{\link{bioFAMmodel}} object
}
\description{
Function to train an untrained \code{\link{bioFAMmodel}} object.
}
\details{
In this step the R package is calling the \code{biofam} Python package, where the the training is performed. \cr
The interface with Python is done with the \code{\link{reticulate}} package. 
If you have several versions of Python installed and Rstudio is not detecting the correct one, you can change it using
\code{reticulate::use_python}. \cr
This module is in beta testing so please, read our FAQ for troubleshooting and report any problems.
}
