% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biofam.R
\name{run_biofam}
\alias{run_biofam}
\title{train a BioFAM model}
\usage{
run_biofam(object, outfile = NA)
}
\arguments{
\item{object}{an untrained \code{\link{bioFAMmodel}} object}

\item{outfile}{output file (use .hdf5 format)}
}
\value{
a trained \code{\link{bioFAMmodel}} object
}
\description{
Function to train an untrained \code{\link{bioFAMmodel}} object.
}
\details{
In this step the R package is calling the \code{biofam} Python package, where the the training is performed. \cr
The interface with Python is done with the \code{\link{reticulate}} package. 
If you have several versions of Python installed and Rstudio is not detecting the correct one, you can change it using
\code{reticulate::use_python}. \cr
This module is in beta testing so please, read our FAQ for troubleshooting and report any problems.
}
