% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{training_data}
\alias{training_data}
\alias{training_data}
\alias{training_data<-}
\alias{training_data,BioFAModel-method}
\title{training_data: set and retrieve training data}
\usage{
training_data(object)

.training_data(object) <- value

\S4method{training_data}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of numeric matrices that contain the training data
}
