% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setMethods.R, R/set_methods.R
\docType{methods}
\name{training_options}
\alias{training_options}
\alias{training_options}
\alias{training_options<-}
\alias{training_options,BioFAModel-method}
\alias{training_options,BioFAModel-method}
\title{training_options: set and retrieve training opts}
\usage{
training_options(object)

.training_options(object) <- value

\S4method{training_options}{BioFAModel}(object)

\S4method{training_options}{BioFAModel}(object)
}
\arguments{
\item{object}{a \code{\link{BioFAModel}} object.}

\item{object}{a \code{\link{BioFAModel}} object.}
}
\value{
list of training options

list of training options
}
